fix(quadraui): drop redundant f32 cast in win draw_toast_stack
`WinBackend::current_line_height` is already `f32`, so the `as f32`
added in the #861 slice tripped `clippy::unnecessary_cast` under the
win examples clippy leg (`-D warnings`, x86_64-pc-windows-msvc).

+++ quadraui/src/win/backend.rs
@@ -2769,7 +2769,7 @@ impl Backend for WinBackend {
         #[cfg(target_os = "windows")]
         if self.surface.is_some() && self.dwrite.is_some() {
             let theme = self.current_theme;
-            let line_height = self.current_line_height as f32;
+            let line_height = self.current_line_height;
             return crate::primitives::toast::native_surface_paint::paint(
                 stack,
                 self,
